backend/requirements.txt
alembic>=1.14.1
apscheduler>=3.10.4
beautifulsoup4>=4.12.3
fastapi>=0.115.7
psycopg2-binary>=2.9.10
pydantic-settings>=2.7.1
python-dateutil>=2.9.0.post0
requests>=2.32.3
sqlalchemy>=2.0.37
uvicorn[standard]>=0.34.0
